Create one root layout to avoid "missing html and body" error in the 404 page
