+++ src/app/layout.tsx
@@ -1,0 +1,22 @@
+import "~/styles/globals.css";
+import { Providers } from "~/app/nav/providers";
+import { GeistSans } from "geist/font/sans";
+import { type Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Brown Puzzle Hunt",
+  description: "",
+  icons: [{ rel: "icon", url: "/favicon.ico" }],
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{ children: React.ReactNode }>) {
+  return (
+    <html lang="en" className={`${GeistSans.variable}`}>
+      <body>
+        {children}
+      </body>
+    </html>
+  );
+}
